Move parsing SV data to its own file
