+++ src/components/LandingPage.vue
@@ -263,7 +263,7 @@ import fs from "fs";
 import got, { Response } from "got";
 import luaparse from "luaparse";
 import path from "path";
-import { defineComponent, reactive } from "vue";
+import { defineComponent } from "vue";
 import { ipcRenderer } from "electron";
 
 import contacts from "@/libs/contacts";
@@ -280,6 +280,7 @@ import {
 import userDataPath from "@/libs/user-data-folder";
 import { matchFolderNameInsensitive, wowDefaultPath } from "@/libs/utilities";
 import { WeakAurasSaved, PlaterSaved } from "@/libs/grab-sv-files";
+import { parseWeakAurasSVdata, parsePlaterSVdata } from "@/libs/parse-sv-data";
 
 import { useStashStore } from "../stores/auras";
 import { useConfigStore } from "../stores/config";
@@ -388,7 +389,7 @@ export default defineComponent({
           addonDependency: "WeakAuras",
           svPathFunction: WeakAurasSaved,
           isInstalled: this.IsAddonInstalled("WeakAuras"),
-          parseFunction: this.parseWeakAurasSVdata,
+          parseFunction: parseWeakAurasSVdata,
           hasTypeColumn: false,
         },
         {
@@ -397,7 +398,7 @@ export default defineComponent({
           addonDependency: "Plater",
           svPathFunction: PlaterSaved,
           isInstalled: this.IsAddonInstalled("Plater"),
-          parseFunction: this.parsePlaterSVdata,
+          parseFunction: parsePlaterSVdata,
           hasTypeColumn: true,
         },
       ];
@@ -673,237 +674,6 @@ export default defineComponent({
           console.log(JSON.stringify(error));
         }
       }
-    },
-    parseWeakAurasSVdata(WeakAurasSavedData, config) {
-      const aurasFromFile = [];
-
-      if (WeakAurasSavedData.body[0].variables[0].name !== "WeakAurasSaved") {
-        return [];
-      }
-
-      const pattern = /(https:\/\/wago.io\/)([^/]+)/;
-
-      WeakAurasSavedData.body[0].init[0].fields.forEach((obj) => {
-        if (obj.key.value === "displays") {
-          obj.value.fields.forEach((obj2) => {
-            let slug;
-            let url;
-            let version = 0;
-            let semver;
-            let ignoreWagoUpdate = false;
-            let id;
-            let uid = null;
-
-            obj2.value.fields.forEach((obj3) => {
-              if (obj3.key === undefined) {
-                return;
-              }
-
-              if (obj3.key.value === "id") {
-                id = obj3.value.value;
-              }
-
-              if (obj3.key.value === "uid") {
-                uid = obj3.value.value;
-              }
-
-              if (obj3.key.value === "version") {
-                version = Number(obj3.value.value);
-              }
-
-              if (obj3.key.value === "semver") {
-                semver = obj3.value.value;
-              }
-
-              if (obj3.key.value === "ignoreWagoUpdate") {
-                ignoreWagoUpdate = obj3.value.value;
-              }
-
-              if (obj3.key.value === "url") {
-                url = obj3.value.value;
-                const result = url.match(pattern);
-
-                if (result) ({ 2: slug } = url.match(pattern));
-              }
-            });
-
-            if (slug) {
-              const foundAura = reactive({
-                id,
-                uid,
-                slug,
-                version,
-                semver,
-                ignoreWagoUpdate,
-                wagoVersion: null,
-                wagoSemver: null,
-                changelog: null,
-                created: null,
-                modified: null,
-                author: null,
-                encoded: null,
-                wagoid: null,
-                ids: [id],
-                uids: uid ? [uid] : [],
-                regionType: null,
-                auraType: config.addonName,
-                auraTypeDisplay: null,
-                addonConfig: config,
-              });
-
-              aurasFromFile.push(foundAura);
-            }
-          });
-        }
-      });
-
-      return aurasFromFile;
-    },
-    parsePlaterSVdata(PlaterSavedData, config) {
-      const aurasFromFile = [];
-
-      if (PlaterSavedData.body[0].variables[0].name !== "PlaterDB") {
-        return;
-      }
-
-      const pattern = /(https:\/\/wago.io\/)([^/]+)/;
-
-      PlaterSavedData.body[0].init[0].fields.forEach((obj) => {
-        if (obj.key.value === "profiles") {
-          obj.value.fields.forEach((profile) => {
-            let profslug;
-            let profurl;
-            let profversion = 0;
-            let profsemver;
-            let profignoreWagoUpdate = false;
-            let profid;
-
-            profile.value.fields.forEach((profData) => {
-              if (profData.key === undefined) {
-                return;
-              }
-
-              if (profData.key.value === "Name") {
-                profid = profData.value.value;
-              }
-
-              if (profData.key.value === "version") {
-                profversion = Number(profData.value.value);
-              }
-
-              if (profData.key.value === "semver") {
-                profsemver = profData.value.value;
-              }
-
-              if (profData.key.value === "ignoreWagoUpdate") {
-                profignoreWagoUpdate = profData.value.value;
-              }
-
-              if (profData.key.value === "url") {
-                profurl = profData.value.value;
-                const result = profurl.match(pattern);
-
-                if (result) ({ 2: profslug } = profurl.match(pattern));
-              }
-
-              if (profData.key.value === "script_data" || profData.key.value === "hook_data") {
-                let typeSuffix =
-                  (profData.key.value === "hook_data" && "-Mod") ||
-                  (profData.key.value === "script_data" && "-Script") ||
-                  "";
-
-                profData.value.fields.forEach((obj2) => {
-                  let slug;
-                  let url;
-                  let version = 0;
-                  let semver;
-                  let ignoreWagoUpdate = false;
-                  let id;
-
-                  obj2.value.fields.forEach((obj3) => {
-                    if (obj3.key.value === "Name") {
-                      id = obj3.value.value;
-                    }
-
-                    if (obj3.key.value === "version") {
-                      version = Number(obj3.value.value);
-                    }
-
-                    if (obj3.key.value === "semver") {
-                      semver = obj3.value.value;
-                    }
-
-                    if (obj3.key.value === "ignoreWagoUpdate") {
-                      ignoreWagoUpdate = obj3.value.value;
-                    }
-
-                    if (obj3.key.value === "url") {
-                      url = obj3.value.value;
-                      const result = url.match(pattern);
-
-                      if (result) ({ 2: slug } = url.match(pattern));
-                    }
-                  });
-
-                  if (slug) {
-                    const foundAura = reactive({
-                      id,
-                      slug,
-                      version,
-                      semver,
-                      ignoreWagoUpdate,
-                      wagoVersion: null,
-                      wagoSemver: null,
-                      changelog: null,
-                      created: null,
-                      modified: null,
-                      author: null,
-                      encoded: null,
-                      wagoid: null,
-                      ids: [id],
-                      uids: [],
-                      regionType: null,
-                      auraType: config.addonName,
-                      auraTypeDisplay: config.addonName + typeSuffix,
-                      addonConfig: config,
-                    });
-
-                    aurasFromFile.push(foundAura);
-                  }
-                });
-              }
-            });
-
-            if (profslug) {
-              const foundAura = reactive({
-                id: profid,
-                slug: profslug,
-                version: profversion,
-                semver: profsemver,
-                ignoreWagoUpdate: profignoreWagoUpdate,
-                wagoVersion: null,
-                wagoSemver: null,
-                changelog: null,
-                created: null,
-                modified: null,
-                author: null,
-                encoded: null,
-                wagoid: null,
-                ids: [profid],
-                uids: [],
-                regionType: null,
-                auraType: config.addonName,
-                auraTypeDisplay: config.addonName + "-Profile",
-                addonConfig: config,
-              });
-
-              aurasFromFile.push(foundAura);
-            }
-          });
-        }
-      });
-
-      return aurasFromFile;
     },
     async compareSVwithWago() {
       if (!this.versionSelected || !this.accountSelected) return;
